% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{from_Clipboard.GeneNames2ID}
\alias{from_Clipboard.GeneNames2ID}
\title{from_Clipboard.GeneNames2ID}
\usage{
from_Clipboard.GeneNames2ID()
}
\description{
Copy genes
}
\examples{
MyVec = c(1:3); # from_Clipboard.GeneNames2ID(MyVec)
}
