% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{fromClipboard.as_named_vec}
\alias{fromClipboard.as_named_vec}
\title{fromClipboard.as_named_vec}
\usage{
fromClipboard.as_named_vec(sep = "\\t", header = F)
}
\arguments{
\item{sep}{PARAM_DESCRIPTION, Default: ' '}

\item{header}{PARAM_DESCRIPTION, Default: F}
}
\description{
Paste a list of strings from your clipboard (e.g. from Excel) into R, parse it to a numeric R vector on OS X.
}
\examples{
# fromClipboard.as_named_vec()
}
