% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{fromClipboard.as_vec}
\alias{fromClipboard.as_vec}
\title{fromClipboard.as_vec}
\usage{
fromClipboard.as_vec(sep = "\\t", header = F)
}
\arguments{
\item{sep}{PARAM_DESCRIPTION, Default: '    '}

\item{header}{PARAM_DESCRIPTION, Default: F}
}
\description{
Paste a list of numbers from your clipboard (e.g. from Excel) into R, parse it to a code-snippet defining an R vector on OS X.
}
