% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{fromClipboard}
\alias{fromClipboard}
\title{fromClipboard}
\usage{
fromClipboard(sep = "\\t", header = F)
}
\arguments{
\item{sep}{PARAM_DESCRIPTION, Default: '    '}

\item{header}{PARAM_DESCRIPTION, Default: F}
}
\description{
Paste data from your clipboard (e.g. a table from Excel) into R, parse it to a code-snippet defining an R data frame on OS X.
}
