% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{inline_list_char}
\alias{inline_list_char}
\title{inline_list_char}
\usage{
inline_list_char(char_list)
}
\arguments{
\item{char_list}{PARAM_DESCRIPTION}
}
\description{
Paste data into your code easily. Take a list of character vectors, parse it to a code-snippet defining an R list, and copy back to the Clipboard.
}
