% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{inline_named_vec.char}
\alias{inline_named_vec.char}
\title{inline_named_vec.char}
\usage{
inline_named_vec.char(num_vector)
}
\arguments{
\item{num_vector}{PARAM_DESCRIPTION}
}
\description{
Take a character vector, parse it to a code-snippet defining an R character vector, and copy back to the Clipboard.
}
