% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{inline_vec.char.from_Clipboard}
\alias{inline_vec.char.from_Clipboard}
\title{inline_vec.char.from_Clipboard}
\usage{
inline_vec.char.from_Clipboard()
}
\description{
Paste data into your code easily. Take a list of strings from your clipboard, parse it to a code-snippet defining an R character vector, and copy back to the Clipboard.
}
\examples{
MyVec = c(1:3); inline_vec.char.from_Clipboard(MyVec)
}
