% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{inline_vec.char}
\alias{inline_vec.char}
\title{inline_vec.char}
\usage{
inline_vec.char(char_vector)
}
\arguments{
\item{char_vector}{PARAM_DESCRIPTION}
}
\description{
Paste data into your code easily. Take a character vector, parse it to a code-snippet defining an R character vector, and copy back to the Clipboard.
}
