% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{inline_vec.num}
\alias{inline_vec.num}
\title{inline_vec.num}
\usage{
inline_vec.num(num_vector)
}
\arguments{
\item{num_vector}{PARAM_DESCRIPTION}
}
\description{
Paste data into your code easily. Take a numeric vector, parse it to a code-snippet defining an R character vector, and copy back to the Clipboard.
}
