% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInCode.R
\name{parse_vec}
\alias{parse_vec}
\title{parse_vec}
\usage{
parse_vec(vector, numeric = F)
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}

\item{numeric}{PARAM_DESCRIPTION, Default: F}
}
\description{
Serialize data. Take a character vector, parse it to a code-snippet defining an R character vector.
}
